Throw exception on attempt to access attr of nonexistent edge object
This is one possible change that could be made to address issue #48.
It also adds many tests for attribute functions, both for existing and
non-existent nodes and edges, in directed and undirected graphs.  For
undirected graphs, it tries variations of accessing both edge
directions, to ensure that they both behave as "the same edge".

Also fix one arity bug.